basic changes to fix the runtime bugs
