+++ backend/tsconfig.json
@@ -14,6 +14,6 @@
     "strict": true,                                      /* Enable all strict type-checking options. */
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
   },
-  "include": ["src/**/*.tsx", "src/index.tsx", "firebase.tsx"],
+  "include": ["src/**/*.tsx", "src/index.ts", "firebase.ts"],
   "exclude": ["node_modules"]
 }
